refactor: extract Parquet schema alignment
